Added likes on posts and categories

+++ blog/views.py
@@ -1,12 +1,14 @@
 from django.shortcuts import render
-from .models import Post
+from .models import Post, Category
 from django.utils import timezone
 from django.shortcuts import render, get_object_or_404
-from .forms import PostForm
+from .forms import PostForm, CategoryForm
 from .forms import SignUpForm
 from django.shortcuts import redirect
 from django.contrib.auth import authenticate, login, logout
 from django.contrib import messages
+from django.http import HttpResponseRedirect
+from django.urls import reverse
 # Create your views here.
 
 def post_list(request):
@@ -28,7 +30,8 @@ def post_list(request):
 
 def post_detail(request, pk):
     post = get_object_or_404(Post, pk = pk)
-    return render(request, 'blog/post_detail.html', {'post':post})
+    total_likes = post.total_likes()
+    return render(request, 'blog/post_detail.html', {'post':post, 'total_likes':total_likes})
 
 def post_new(request):
     if request.method == "POST":
@@ -66,8 +69,12 @@ def post_publish(request, pk):
 
 def post_remove(request, pk):
     post = get_object_or_404(Post, pk=pk)
-    post.delete()
-    return redirect('post_list')
+
+    if request.method == 'POST':
+        post.delete()
+        return redirect('post_list')
+
+    return render(request, 'delete_post.html', {'post': post})
 
 def logout_user(request):
     logout(request)
@@ -91,3 +98,25 @@ def register_user(request):
         return render(request, 'register.html', {'form':form})
 
     return render(request, 'register.html', {'form':form})
+
+def AddCategoryView(request):
+    if request.method == 'POST':
+        form = CategoryForm(request.POST)
+        if form.is_valid():
+            # Save the form data to the database or perform any other necessary actions
+            form.save()
+            # Redirect to a success page or any other desired page
+            return redirect('post_list')  # Replace 'success_url_name' with the actual name of your success URL
+
+    else:
+        form = CategoryForm()
+    return render(request, 'add_category.html', {'form': form})
+
+def category_views(request, cats):
+    category_posts = Post.objects.filter(category=cats.replace('-', ' '))
+    return render(request, 'categories.html', {'cats':cats.title().replace('-', ' '), 'category_posts':category_posts})
+
+def LikeView(request, pk):
+    post = get_object_or_404(Post, id = request.POST.get('post_id'))
+    post.likes.add(request.user)
+    return HttpResponseRedirect(reverse('post_detail', args=[str(pk)]))
